server/requirements.txt
APScheduler==3.11.2
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
chardet==3.0.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
deep-translator==1.11.4
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.2
fsspec==2026.1.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.36.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
mysql-connector-python==9.6.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
protobuf==3.20.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rfc3986==1.5.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.5